test: add failing test for validation of lotto numbers
- test count of lotto numbers
- test range of lotto numbers
- test duplicated lotto numbers
