+++ README.md
@@ -101,9 +101,9 @@
 
 ### 1. 로또 번호 기능 테스트
 - [ ] 로또 번호 검증 기능을 테스트한다.
-    - [ ] `예외사항`: 로또 번호가 6개의 숫자가 아니라면 예외 발생
-    - [ ] `예외사항`: 로또 번호가 1~45 사이의 숫자가 아니라면 예외 발생
-    - [ ] `예외사항`: 로또 번호에 중복된 숫자가 존재하면 예외 발생
+    - ❌ `예외사항`: 로또 번호가 6개의 숫자가 아니라면 예외 발생
+    - ❌ `예외사항`: 로또 번호가 1~45 사이의 숫자가 아니라면 예외 발생
+    - ❌ `예외사항`: 로또 번호에 중복된 숫자가 존재하면 예외 발생
 - [ ] 로또 번호이 오름차순으로 정렬되었는지 테스트한다.
 
 ### 2. 로또 번호 생성 기능 테스트
